+++ CHANGELOG.md
@@ -15,10 +15,13 @@
 - Added `isConvenienceInitializer` property for `Method`
 - Added `defaultValue` for variables and method parameters
 - Added docs generated with jazzy
+- Sourcery now will not create empty files and will remove existing generated files with empty content if CLI flag `prune` is set to `true` (`false` by default)
+- Sourcery now will remove inline annotation comments from generated code.
 
 ### Fixed
 
 - Inserting multiple inline code block in one file
+- Suppress warnings when compiling swift templates
 - Accessing protocols in Swift templates
 
 ## 0.5.9